# Use native JSI Runtime evaluate func