--- a/Common/cpp/Registries/WorkletsCache.cpp
+++ b/Common/cpp/Registries/WorkletsCache.cpp
@@ -7,6 +7,7 @@ using namespace facebook;
 namespace reanimated
 {
 
+// TODO: Those functions never get released, memory leak?
 static std::unordered_map<long long, std::shared_ptr<jsi::Function>> worklets;
 
 jsi::Value eval(jsi::Runtime &rt, const char *code) {