--- a/Common/cpp/SharedItems/FrozenObject.cpp
+++ b/Common/cpp/SharedItems/FrozenObject.cpp
@@ -4,12 +4,12 @@
 
 namespace reanimated {
 
-FrozenObject::FrozenObject(jsi::Runtime &rt, const jsi::Object &object, std::shared_ptr<ErrorHandler> errorHandler, std::shared_ptr<Scheduler> uiScheduler, std::shared_ptr<ShareableValue> valueSetter) {
+FrozenObject::FrozenObject(jsi::Runtime &rt, const jsi::Object &object, NativeReanimatedModule *module) {
   auto propertyNames = object.getPropertyNames(rt);
   for (size_t i = 0, count = propertyNames.size(rt); i < count; i++) {
     auto propertyName = propertyNames.getValueAtIndex(rt, i).asString(rt);
     std::string nameStr = propertyName.utf8(rt);
-    map[nameStr] = ShareableValue::adapt(rt, object.getProperty(rt, propertyName), errorHandler, uiScheduler, valueSetter);
+    map[nameStr] = ShareableValue::adapt(rt, object.getProperty(rt, propertyName), module);
     this->containsHostFunction |= map[nameStr]->containsHostFunction;
   }
 }