--- a/Common/cpp/SharedItems/RemoteObject.cpp
+++ b/Common/cpp/SharedItems/RemoteObject.cpp
@@ -1,6 +1,7 @@
 #include "RemoteObject.h"
 #include "SharedParent.h"
 #include "NativeReanimatedModule.h"
+#include "RuntimeDecorator.h"
 #include <jsi/jsi.h>
 
 using namespace facebook;
@@ -16,14 +17,14 @@ void RemoteObject::maybeInitializeOnUIRuntime(jsi::Runtime &rt) {
 }
 
 jsi::Value RemoteObject::get(jsi::Runtime &rt, const jsi::PropNameID &name) {
-  if (module->isUIRuntime(rt)) {
+  if (RuntimeDecorator::isUIRuntime(rt)) {
     return backing.lock()->getObject(rt).getProperty(rt, name);
   }
   return jsi::Value::undefined();
 }
 
 void RemoteObject::set(jsi::Runtime &rt, const jsi::PropNameID &name, const jsi::Value &value) {
-  if (module->isUIRuntime(rt)) {
+  if (RuntimeDecorator::isUIRuntime(rt)) {
     backing.lock()->getObject(rt).setProperty(rt, name, value);
   }
   // TODO: we should throw if trying to update remote from host runtime