--- a/Common/cpp/SharedItems/RemoteObject.cpp
+++ b/Common/cpp/SharedItems/RemoteObject.cpp
@@ -1,6 +1,5 @@
 #include "RemoteObject.h"
 #include "SharedParent.h"
-#include "NativeReanimatedModule.h"
 #include "RuntimeDecorator.h"
 #include <jsi/jsi.h>