--- a/Common/cpp/SharedItems/RemoteObject.cpp
+++ b/Common/cpp/SharedItems/RemoteObject.cpp
@@ -16,14 +16,14 @@ void RemoteObject::maybeInitializeOnUIRuntime(jsi::Runtime &rt) {
 }
 
 jsi::Value RemoteObject::get(jsi::Runtime &rt, const jsi::PropNameID &name) {
-  if (RuntimeDecorator::isUIRuntime(rt)) {
+  if (RuntimeDecorator::isWorkletRuntime(rt)) {
     return backing.lock()->getObject(rt).getProperty(rt, name);
   }
   return jsi::Value::undefined();
 }
 
 void RemoteObject::set(jsi::Runtime &rt, const jsi::PropNameID &name, const jsi::Value &value) {
-  if (RuntimeDecorator::isUIRuntime(rt)) {
+  if (RuntimeDecorator::isWorkletRuntime(rt)) {
     backing.lock()->getObject(rt).setProperty(rt, name, value);
   }
   // TODO: we should throw if trying to update remote from host runtime