--- a/Common/cpp/headers/Tools/RuntimeDecorator.h
+++ b/Common/cpp/headers/Tools/RuntimeDecorator.h
@@ -20,6 +20,7 @@ public:
                                 TimeProviderFunction getCurrentTime);
   
   static bool isUIRuntime(jsi::Runtime &rt);
+  static bool isReactRuntime(jsi::Runtime &rt);
 };
 
 }