… Common/cpp/Tools/Mapper.cpp
… Common/cpp/Tools/RuntimeDecorator.cpp
… Common/cpp/Tools/Scheduler.cpp
… Common/cpp/headers/NativeModules/NativeReanimatedModule.h
… Common/cpp/headers/Registries/WorkletsCache.h
… Common/cpp/headers/SharedItems/FrozenObject.h
… Common/cpp/headers/SharedItems/MutableValue.h
… Common/cpp/headers/SharedItems/RemoteObject.h
… Common/cpp/headers/SharedItems/RuntimeManager.h
… Common/cpp/headers/SharedItems/ShareableValue.h
… Common/cpp/headers/SpecTools/ErrorHandler.h
… Common/cpp/headers/Tools/Mapper.h
… Common/cpp/headers/Tools/Scheduler.h
… Example/android/gradle.properties
… Example/ios/Podfile.lock
… Example/ios/ReanimatedExample.xcodeproj/project.pbxproj
… Example/package.json
… Example/yarn.lock
… RNReanimated.podspec
… android/src/main/cpp/NativeProxy.cpp
… ios/RNReanimated.xcodeproj/project.pbxproj
… ios/native/NativeProxy.mm
… ios/native/REAIOSScheduler.mm
… ios/native/UIResponder+Reanimated.mm
… plugin.js