## Description

This is a follow-up of https://github.com/software-mansion/react-native-reanimated/pull/1790.

It is purely a code quality improvement, as the result is roughly the same.

## Changes

* Replace the retrieval of global.eval(...), calling it, and getting the returned function with a simple `Runtime::evaluateJavaScript(...)`

I have benchmarked this but couldn't notice any significant changes, so I assume it is basically doing the same. I also tried to use `prepareJavaScript` and then `evaluatePreparedJavaScript` but that also didn't change anything.

## Benchmarks

### Before:

```
2021-03-18 09:14:12.946440+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 56ms
2021-03-18 09:14:12.947215+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 68ms
2021-03-18 09:14:12.947460+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 51ms
2021-03-18 09:14:12.947821+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 86ms
2021-03-18 09:14:12.947999+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 19ms
2021-03-18 09:14:12.948240+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 49ms
2021-03-18 09:14:12.948516+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 31ms
2021-03-18 09:14:12.948730+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 37ms
2021-03-18 09:14:12.949116+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 50ms
2021-03-18 09:14:12.949329+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 40ms
2021-03-18 09:14:12.949512+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 21ms
2021-03-18 09:14:12.949845+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 147ms
2021-03-18 09:14:12.950113+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 54ms
2021-03-18 09:14:12.950468+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 126ms
2021-03-18 09:14:12.950890+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 49ms
2021-03-18 09:14:12.951242+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 63ms
2021-03-18 09:14:12.951637+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 21ms
2021-03-18 09:14:12.951850+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 18ms
2021-03-18 09:14:12.952636+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 17ms
2021-03-18 09:14:12.952805+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 19ms
2021-03-18 09:14:12.952964+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 20ms
2021-03-18 09:14:12.953175+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 24ms
2021-03-18 09:14:12.953382+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 23ms
2021-03-18 09:14:12.953557+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 19ms
2021-03-18 09:14:12.953706+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 23ms
2021-03-18 09:14:12.954333+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 57ms
2021-03-18 09:14:12.954695+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 38ms
2021-03-18 09:14:12.954908+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 57ms
2021-03-18 09:14:12.955136+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 58ms
2021-03-18 09:14:12.955427+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 59ms
2021-03-18 09:14:12.955612+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 48ms
2021-03-18 09:14:12.955995+0100 ReanimatedExample[9344:189303] Evaluating Worklet took 20ms
```

### After

```
2021-03-18 09:14:42.995224+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 54ms
2021-03-18 09:14:42.995998+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 59ms
2021-03-18 09:14:42.996217+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 42ms
2021-03-18 09:14:42.996516+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 46ms
2021-03-18 09:14:42.996696+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 17ms
2021-03-18 09:14:42.996898+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 43ms
2021-03-18 09:14:42.997110+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 16ms
2021-03-18 09:14:42.997349+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 41ms
2021-03-18 09:14:42.997713+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 43ms
2021-03-18 09:14:42.997960+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 39ms
2021-03-18 09:14:42.998144+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 15ms
2021-03-18 09:14:42.998422+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 99ms
2021-03-18 09:14:42.998655+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 49ms
2021-03-18 09:14:42.998978+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 117ms
2021-03-18 09:14:42.999374+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 52ms
2021-03-18 09:14:42.999786+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 35ms
2021-03-18 09:14:43.000217+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 17ms
2021-03-18 09:14:43.000452+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 18ms
2021-03-18 09:14:43.001388+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 15ms
2021-03-18 09:14:43.001571+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 15ms
2021-03-18 09:14:43.001734+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 17ms
2021-03-18 09:14:43.001910+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 17ms
2021-03-18 09:14:43.002085+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 18ms
2021-03-18 09:14:43.002272+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 16ms
2021-03-18 09:14:43.002458+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 20ms
2021-03-18 09:14:43.003233+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 61ms
2021-03-18 09:14:43.003608+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 35ms
2021-03-18 09:14:43.003878+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 63ms
2021-03-18 09:14:43.004151+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 56ms
2021-03-18 09:14:43.004401+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 54ms
2021-03-18 09:14:43.004579+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 40ms
2021-03-18 09:14:43.004937+0100 ReanimatedExample[9361:190498] Evaluating Worklet took 13ms
```

> ms = microseconds