@mrousavy 2021-03-23
Closing in favor of https://github.com/software-mansion/react-native-reanimated/pull/1863.